Move to the localization namespace
