+++ src/OrchardCore/OrchardCore/Localization/CalendarSelectorResult.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Threading.Tasks;
 
-namespace OrchardCore.Modules
+namespace OrchardCore.Localization
 {
     public class CalendarSelectorResult
     {
