+++ src/OrchardCore/OrchardCore/Localization/ICalendarSelector.cs
@@ -1,7 +1,6 @@
-using System;
 using System.Threading.Tasks;
 
-namespace OrchardCore.Modules
+namespace OrchardCore.Localization
 {
     public interface ICalendarSelector
     {
